refactor: add packages and move classes to new package
