+++ src/main/java/baseball/controller/GameController.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.controller;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
@@ -11,7 +11,7 @@ public class GameController {
     private final String TEXT_RESULT_STRIKE = "스트라이크";
     private final String TEXT_RESULT_BALL = "볼 ";
 
-    ArrayList<Integer> getAnswer() {
+    public ArrayList<Integer> getAnswer() {
         // 중복 제거를 위한 HashSet
         HashSet<Integer> set = new HashSet<>(NUMBER_COUNT);
         while (set.size() < NUMBER_COUNT) {
@@ -20,7 +20,7 @@ public class GameController {
         return new ArrayList<>(set);
     }
 
-    List<Integer> getInputFromString(String playerInput) {
+    public List<Integer> getInputFromString(String playerInput) {
         List<Integer> input = new ArrayList<>();
         for (int i = 0; i < playerInput.length(); i++) {
             input.add(playerInput.charAt(i) - '0');
@@ -28,7 +28,7 @@ public class GameController {
         return input;
     }
 
-    int countStrike(List<Integer> answer, List<Integer> input) {
+    public int countStrike(List<Integer> answer, List<Integer> input) {
         int count = 0;
         for (int i = 0; i < NUMBER_COUNT; i++) {
             count = count + checkStrike(answer.get(i), input.get(i));
@@ -43,7 +43,7 @@ public class GameController {
         return 0;
     }
 
-    int countBall(List<Integer> answer, List<Integer> input) {
+    public int countBall(List<Integer> answer, List<Integer> input) {
         int count = 0;
         for (int i = 0; i < NUMBER_COUNT; i++) {
             count = count + checkBall(answer, answer.get(i), input.get(i));
@@ -58,7 +58,7 @@ public class GameController {
         return 0;
     }
 
-    String getHint(int strike, int ball) {
+    public String getHint(int strike, int ball) {
         return getBallMessage(ball) + getStrikeMessage(strike);
     }
