+++ src/main/java/baseball/model/GameAnswerModel.java
@@ -1,0 +1,5 @@
+package baseball.model;
+
+public class GameAnswerModel {
+
+}
