+++ src/main/java/baseball/model/GameInputModel.java
@@ -1,0 +1,5 @@
+package baseball.model;
+
+public class GameInputModel {
+
+}
